+++ config/module.config.php
@@ -40,6 +40,9 @@
 
 namespace Soflomo\Purifier;
 
+use HTMLPurifier;
+use HTMLPurifier_Config;
+
 return array(
     'soflomo_purifier' => [
         'standalone'      => false,
@@ -50,7 +53,8 @@ return array(
 
     'service_manager' => [
         'factories' => [
-            'HTMLPurifier' => Factory\HtmlPurifierFactory::class,
+            HTMLPurifier::class        => Factory\HtmlPurifierFactory::class,
+            HTMLPurifier_Config::class => Factory\HtmlPurifierConfigFactory::class,
         ],
     ],