+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Support for HTMLPurifier custom definitions ([#1](https://github.com/juriansluiman/Soflomo-Purifier/pull/1))
+- The default `HTMLPurifier_Config` instance is now available as a service via its FQCN.
 
 ### Changed
 - Updated the docs in README.md